+++ SmartPower/UI/SmartGoodPoweredGeneratorFragment.cs
@@ -3,6 +3,7 @@
 // License: Public Domain
 
 using IgorZ.TimberDev.UI;
+using TimberApi.UIPresets.Buttons;
 using Timberborn.BaseComponentSystem;
 using Timberborn.CoreUI;
 using Timberborn.EntityPanelSystem;
@@ -17,11 +18,16 @@ sealed class SmartGoodPoweredGeneratorFragment : IEntityPanelFragment {
   const string NeverStopThisGeneratorLocKey = "IgorZ.SmartPower.PoweredGenerator.NeverStop";
   const string ChargeLevelLocKey = "IgorZ.SmartPower.PoweredGenerator.ChargeBatteriesRangeText";
 
+  static readonly Vector2 LessChargeRange = new(0.30f, 0.55f);
+  static readonly Vector2 MoreChargeRange = new(0.65f, 0.90f);
+
   readonly UiFactory _uiFactory;
 
   VisualElement _root;
   Toggle _neverShutdownCheckbox;
   Label _chargeBatteriesText;
+  Button _lessChargeButton;
+  Button _moreChargeButton;
   MinMaxSlider _chargeBatteriesSlider;
   SmartGoodPoweredGenerator _generator;
 
@@ -41,15 +47,33 @@ sealed class SmartGoodPoweredGeneratorFragment : IEntityPanelFragment {
           _generator.DischargeBatteriesThreshold = evt.newValue.x;
           _generator.ChargeBatteriesThreshold = evt.newValue.y;
           UpdateControls();
-        }, 0f, 1.0f, 0.10f, stepSize: 0.01f);
+        }, 0f, 1.0f, 0.10f, stepSize: 0.05f);
+    _chargeBatteriesSlider.style.marginLeft = 10;
+    _chargeBatteriesSlider.style.marginRight = 10;
 
     _chargeBatteriesText = _uiFactory.CreateLabel();
     _chargeBatteriesText.style.marginTop = 5;
+    _chargeBatteriesText.style.marginBottom = 5;
+    
+    _lessChargeButton = _uiFactory.UiBuilder.Create<ArrowLeftButton>().Build();
+    _lessChargeButton.clicked += () => {
+      _chargeBatteriesSlider.value = LessChargeRange;
+    };
+    _moreChargeButton = _uiFactory.UiBuilder.Create<ArrowRightButton>().Build();
+    _moreChargeButton.clicked += () => {
+      _chargeBatteriesSlider.value = MoreChargeRange;
+    };
+
+    var panel = new VisualElement();
+    panel.style.flexDirection = FlexDirection.Row;
+    panel.Add(_lessChargeButton);
+    panel.Add(_chargeBatteriesSlider);
+    panel.Add(_moreChargeButton);
 
     _root = _uiFactory.CreateCenteredPanelFragmentBuilder()
         .AddComponent(_neverShutdownCheckbox)
         .AddComponent(_chargeBatteriesText)
-        .AddComponent(_chargeBatteriesSlider)
+        .AddComponent(panel)
         .BuildAndInitialize();
 
     _root.ToggleDisplayStyle(visible: false);
@@ -79,6 +103,8 @@ sealed class SmartGoodPoweredGeneratorFragment : IEntityPanelFragment {
     _chargeBatteriesText.text = _uiFactory.Loc.T(
         ChargeLevelLocKey, Mathf.RoundToInt(_generator.DischargeBatteriesThreshold * 100),
         Mathf.RoundToInt(_generator.ChargeBatteriesThreshold * 100));
+    _moreChargeButton.SetEnabled(_chargeBatteriesSlider.value != MoreChargeRange);
+    _lessChargeButton.SetEnabled(_chargeBatteriesSlider.value != LessChargeRange);
   }
 }