+++ SmartPower/CHANGELOG.md
@@ -1,5 +1,7 @@
 # v1.9.1 (15 Aug 2024)
 * [Change] Idle attractions now consume 1hp. It solves the edge case with the unpowered networks.
+* [Update] Set 5% steps for the charge battery slider.
+* [Update] Add "less charge" / "more charge" buttons to quickly set slider presets.
 
 # v1.9.0 (preview)
 * [Update] Change `deDE` strings.